requirements.txt
flask
ultralytics
opencv-python
numpy
pillow
pytesseract
